🚚 (Sidebar) - Move sidebar.ts in app folder
